#311 NoteService를 lifecycle/share/reference-rewrite 서비스로 분리
48KB 단일 클래스였던 NoteService의 책임을 세 개로 분리:
- NoteLifecycleService: archive/unarchive, 휴지통(soft delete/restore),
  영구 삭제(단건/비우기/보존기간 purge) 및 관련 재귀 CTE 헬퍼
- NoteShareService: 공유 토큰 발급/해제, 익명 토큰 조회, 콘텐츠 재-sanitize 백필
- NoteReferenceRewriter: 페이지링크/멘션 제목 전파·제거 등 순수 콘텐츠 재작성,
  참조 스캔 프로브(WhereReferencesNote), StripHtml, 쓰기 타임스탬프 정규화

NoteService는 CRUD/검색/백링크/제목 전파 핵심 로직만 유지하고 lifecycle/share
메서드는 주입받은 두 서비스로 위임하여 컨트롤러·테스트 표면(public API)을 그대로 보존.
IgnoreQueryFilters() 사용 지점과 아카이브 가시성 규칙은 원본 그대로 이전.

Program.cs에 두 서비스를 Scoped로 등록, TestHost도 동일하게 배선.
NoteReferenceRewriter 순수 헬퍼에 대한 직접 단위 테스트 추가.
